+++ src/update-generator/update-generator.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { UpdateGeneratorService } from './update-generator.service';
+import { UpdateGeneratorController } from './update-generator.controller';
+
+@Module({
+  controllers: [UpdateGeneratorController],
+  providers: [UpdateGeneratorService],
+})
+export class UpdateGeneratorModule {}